refactor: separating presentation from data, at least a little
