--- src/data/releases/mod.rs
@@ -1,1 +1,0 @@
-pub mod github;
