+++ src/site/artifacts/package_managers.rs
@@ -1,6 +1,6 @@
 use crate::config::Config;
-use crate::data::artifacts::get_copyicon;
 use crate::errors::*;
+use crate::site::icons;
 use crate::site::markdown::{syntax_highlight, SyntaxTheme};
 use linked_hash_map::LinkedHashMap;
 
@@ -14,20 +14,20 @@ fn create_package_install_code(code: &str, syntax_theme: &SyntaxTheme) -> String
         Err(_) => format!("<code class='text-center break-all'>{}</code>", code),
     }
 }
+
 // False positive duplicate allocation warning
 // https://github.com/rust-lang/rust-clippy/issues?q=is%3Aissue+redundant_allocation+sort%3Aupdated-desc
 #[allow(clippy::vec_box)]
 pub fn build_list(managers: &LinkedHashMap<String, String>, config: &Config) -> Box<div<String>> {
     let mut list = vec![];
     for (manager, install_code) in managers.iter() {
-        let copy_icon = get_copyicon();
         list.extend(html!(<li class="list-none"><h5>{text!(manager)}</h5> 
         <div class="install-code-wrapper">
         {unsafe_text!(create_package_install_code(install_code, &config.syntax_theme))}
         <button
             data-copy={install_code}
             class="button primary button">
-            {copy_icon}
+            {icons::copy()}
         </button>
     </div>
         
@@ -45,7 +45,7 @@ pub fn build_list(managers: &LinkedHashMap<String, String>, config: &Config) -> 
     )
 }
 
-pub fn build(
+pub fn build_header(
     config: &Config,
     package_managers: &LinkedHashMap<String, String>,
 ) -> Result<Box<div<String>>> {
